Mark the need to use `readFile` using only a footnote.
While `node` and `bun` could support `fetch` for files (and there are talks to do so), it is arguably a security concern to allow `fetch` to work on `file://` URLs by default. It is also not unreasonable to treat local filesystem access as a different primitive from URL fetching, as there are differences in functionality, implementation, performance, etc.

So I think it's fair to consider it a reasonable level of support if `readFile(…)` works. There are still some subtle quirks in how how `file://` URLs are created and support (both as strings and `URL` objects), but those are not sufficient enough to get in the way of writing direct, cross-compatible code — assuming you test against all the platforms.